feat: move vehicle-recall banner to a component
