+++ src/components/common/common-components.module.ts
@@ -32,6 +32,7 @@ import { ChartComponent } from '@components/common/chart/chart';
 import { DataGridComponent } from '@components/common/data-grid/data-grid';
 import { LogoutModalModule } from '@components/common/logout-modal/logout-modal.module';
 import { TestRecoveredBannerComponent } from '@components/common/test-recovered-banner/test-recovered-banner';
+import { VehicleRecallsBanner } from '@components/common/vehicle-recall-banner/vehicle-recalls-banner';
 import { DirectivesModule } from '@directives/directives.module';
 import { NgApexchartsModule } from 'ng-apexcharts';
 import { BikeCategoryTypeComponent } from './bike-category-type/bike-category-type';
@@ -63,6 +64,7 @@ import { WarningBannerComponent } from './warning-banner/warning-banner';
     EndTestLinkComponent,
     ErrorMessageComponent,
     OfflineBannerComponent,
+    VehicleRecallsBanner,
     TabComponent,
     TabsComponent,
     WarningBannerComponent,
@@ -117,6 +119,7 @@ import { WarningBannerComponent } from './warning-banner/warning-banner';
     EndTestLinkComponent,
     ErrorMessageComponent,
     OfflineBannerComponent,
+    VehicleRecallsBanner,
     TabComponent,
     TabsComponent,
     WarningBannerComponent,
