+++ src/components/common/vehicle-recall-banner/vehicle-recalls-banner.scss
@@ -1,0 +1,9 @@
+.stop-drive-banner {
+  background-color: #16eeef
+}
+
+.warning-icon {
+  padding-right: 5px;
+  font-size: 58px;
+  color: black;
+}
